Refactor datasetprofile's flatten profile (#250)
* add missing type

removed unused variable

format fix

* refactor flatten summary functions

* refactor flatten summary from datasetprofile

* add more unit tests.

* remove type conversion for strings

* build with poetry for release

* revert commit for type change

* revert string type commit
